% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adiff.R
\name{adiff}
\alias{adiff}
\title{Automatic Differentiation Using Madness}
\usage{
adiff(func, x, ...)
}
\arguments{
\item{func}{The function to be differentiated (usually user-defined).}

\item{x}{The values at which to evaluate the function \code{func}.}

\item{...}{Further arguments to be passed to \code{func}.}
}
\value{
A numeric vector or matrix containing the result of the function 
  evaluation \code{func(x, ...)}. The \code{"gradient"} attribute of this 
  object contains the Jacobian matrix of the transformation \code{func}.
}
\description{
A wrapper function to the functionality of the \code{madness} package. 
Converts a given \code{x} to a madness object and then applies \code{func} to
it, returning the result and the Jacobian for the transformation \code{func}.
\code{adiff} is used by the \code{\link{rjmcmcpost}} function.
}
\examples{
x2x3 = function(x){
return(c(x^2, x^3))
}

y = adiff(x2x3, c(5,6))
attr(y, "gradient")

}
\references{
Pav, S. E. (2016) Madness: Automatic Differentiation of
  Multivariate Operations.
}
\seealso{
\code{\link{madness}}
}
