% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rjmcmcMethodsIntern.R
\name{birthMoveK1}
\alias{birthMoveK1}
\title{Birth move in the case that only one nucleosome is present}
\usage{
birthMoveK1(paramValues, kValue, muValue, sigmafValue, sigmarValue, deltaValue,
  wValue, dlValue, aValue, dimValue)
}
\arguments{
\item{paramValues}{= list(startPSF =  startPosForwardReads, startPSR = startPosReverseReads
   , kmax = kmax, lambda = lambda, minReads = minReads
   , y = y, nr = nr, nf =nf, nbrReads = nbrReads , zeta = zeta
   , deltamin = deltamin, deltamax = deltamax
   , minReadPos= minReadPos, maxReadPos = maxReadPos )}

\item{kValue}{}

\item{muValue}{}

\item{sigmafValue}{}

\item{sigmarValue}{}

\item{deltaValue}{}

\item{wValue}{}

\item{dlValue}{}

\item{aValue}{}

\item{dimValue}{}
}
\value{
list( k=0L, mu=numeric(paramValues$kmax)
   , sigmaf=numeric(paramValues$kmax), sigmar=numeric(paramValues$kmax)
   , delta=numeric(paramValues$kmax), w=numeric(paramValues$kmax)
   , dl=numeric(paramValues$kmax), a=numeric(paramValues$kmax)
   , dim=numeric(paramValues$kmax), rho=0)
}
\description{
Attempts to add a component for a new nucleosome in the
case that only one nucleosome is present \code{k = 1}
}
\examples{
birthMoveK1(paramValues, kValue, muValue, sigmafValue, sigmarValue,
deltaValue, wValue, dlValue, aValue, dimValue )
}
\author{
Rawane Samb, Pascal Belleau, Astrid Deschenes
}
\keyword{internal}

