% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rjmcmcMethodsIntern.R
\name{Dk}
\alias{Dk}
\title{Death Submove Probability}
\usage{
Dk(k, lambda, kMax)
}
\arguments{
\item{k}{a \code{numeric} which is the number of nucleosomes.}

\item{lambda}{a \code{numeric} used as the lambda
     value in the truncated Poisson distribution.}

\item{kMax}{a \code{numeric} indicating the maximum
     value authorized for the \code{k} parameter. When
     \code{k} is equal or superior to \code{kMax}, the
     returned value is \code{0}.}
}
\value{
a \code{numeric} value. The value \code{0} when
     \code{k} is equal or superior to \code{kMax} or
     when \code{k} is equal to \code{1}.
}
\description{
Calculation of the death submove
     probability pf a randomly selected nucleosome using
     a truncated Poisson distribution.
}
\examples{
## Return the death submove probability
rjmcmc:::Dk(k = 12L, lambda = 4.4, kMax = 30L)

## Zero is returned when k = 1
rjmcmc:::Dk(k = 1L, lambda = 4.4, kMax = 30L)

## Zerio is returned when k is superior to kMax
rjmcmc:::Dk(k = 31L, lambda = 4.4, kMax = 30L)
}
\author{
Rawane Samb
}
\keyword{internal}

