% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rjmcmcMethodsIntern.R
\name{elementWithHighestMode}
\alias{elementWithHighestMode}
\title{Element with the hightest number of occurences}
\usage{
elementWithHighestMode(sample)
}
\arguments{
\item{sample}{a \code{numeric} \code{vector} (of positive \code{integer}
values). If the elements of \code{sample} are \code{numeric} but not
\code{integer}, the elements are truncated by \code{as.integer}.}
}
\value{
a \code{integer} with the highest number of occurences or
\code{NA} when more than one \code{integer} have the highest number
of occurences.
}
\description{
Returned the \code{integer} with the highest number of
occurences in a \code{vector}.
When more than one \code{integer} have the highest number of occurences,
\code{NA} is returned.
}
\examples{
## Return the element with the hightest number of occurence
data01 <- c(1L, 2L, 5L, 10L, 5L, 10L, 5L)
rjmcmc:::elementWithHighestMode(data01)

data02 <- c(3L, 6L, 4L, 3L, 6L)
rjmcmc:::elementWithHighestMode(data02)
}
\author{
Rawane Samb, Astrid Deschenes
}
\keyword{internal}

