% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goals.R
\docType{data}
\name{goals}
\alias{goals}
\title{Football goals scored}
\format{A vector of length 1140, filled with integers.}
\source{
www.football-data.co.uk
}
\usage{
goals
}
\description{
Count data regarding the total number of goals scored in each match over three seasons of the English Premier League football competition, from 2005/06 to 2007/08.
}
\keyword{datasets}
