% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rjmcmcMethodsIntern.R
\name{isInteger}
\alias{isInteger}
\title{Validate if a value is an integer}
\usage{
isInteger(value)
}
\arguments{
\item{value}{an object to validate.}
}
\value{
\code{TRUE} is the parameter is a integer; otherwise \code{FALSE}
}
\description{
Validate if the value passed to the function can be casted
into a \code{integer} or
not. The value must have a length
of 1. The type of value can be a \code{integer} or \code{numerical}.
}
\examples{
## Return TRUE because the input is an integer of length 1
rjmcmc:::isInteger(33L)

## Return FALSE because the length of the input is not 1
rjmcmc:::isInteger(c(21L, 33L))

## Return TRUE because the input is a numericalof length 1
rjmcmc:::isInteger(323.21)

## Return FALSE because the length of the input is not 1
rjmcmc:::isInteger(c(444.2, 442.1))
}
\author{
Astrid Louise Deschenes
}
\keyword{internal}

