% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjmcmc.R
\name{kernel_ftn}
\alias{kernel_ftn}
\title{kernel function for something something}
\usage{
kernel_ftn(t_input, l_input, g_input)
}
\arguments{
\item{t_input}{time vector}

\item{l_input}{range parameter}

\item{g_input}{nugget}
}
\value{
calculated kernel matrix length(t_input) x length(t_input)
}
\description{
calculating the gaussian  kernel  used in rjmcmc
}
