% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjmcmc.R
\name{log_zip_lkh}
\alias{log_zip_lkh}
\title{zero-inflated Poisson likelihood}
\usage{
log_zip_lkh(y_input, pi_input, lambda_input)
}
\arguments{
\item{y_input}{observation vector}

\item{pi_input}{zero-probability}

\item{lambda_input}{poisson intensity}
}
\value{
calculated log likelihood (scalar)
}
\description{
calculating the log likelihood of zip used in rjmcmc
}
\examples{
log_zip_lkh(c(rep(0, 5), rpois(10, 1)), 0.2, 1)

}
