% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rjmcmcMethodsIntern.R
\name{mergeAllRDSFiles}
\alias{mergeAllRDSFiles}
\title{Merge nucleosome information}
\usage{
mergeAllRDSFiles(arrayOfFiles)
}
\arguments{
\item{arrayOfFiles}{a \code{array}, the name of each file that must be
used to merge nucleosome information.}
}
\value{
a \code{list} of \code{class} "rjmcmcNucleosomesMerge" containing:
\itemize{
    \item k a \code{integer}, the number of nucleosomes.
    \item mu a \code{vector} of \code{numeric} of length
\code{k}, the positions of the nucleosomes.
    \item sigmaf a \code{vector} of \code{numeric} of length
\code{k}, the variance of the forward reads for each nucleosome.
    \item sigmar a \code{vector} of \code{numeric} of length
\code{k}, the variance of the reverse reads for each nucleosome.
    \item delta a \code{vector} of \code{numeric} of length
\code{k}, the distance between the maxima of the forward and
reverse reads position densities for each nucleosome.
}
}
\description{
Merge nucleosome information present in multiple RDS files.
}
\examples{
## TODO
}
\author{
Pascal Belleau, Astrid Deschenes
}
\keyword{internal}

