% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rjmcmcMethodsIntern.R
\name{mergeNucleosomes}
\alias{mergeNucleosomes}
\title{Merging two nucleosomal regions}
\usage{
mergeNucleosomes(yf, yr, y, liste, ecartmin, ecartmax, minReads)
}
\arguments{
\item{yf}{a \code{vector} of positive \code{integer}
corresponding to the
positions of all forward reads. The
values insinde \code{yf} must be sorted.}

\item{yr}{a \code{vector} of positive \code{integer}
corresponding to the
positions of all reverse reads. The
values insinde \code{yf} must be sorted.}

\item{y}{TODO}

\item{liste}{TODO}

\item{ecartmin}{TODO}

\item{ecartmax}{TODO}

\item{minReads}{TODO}
}
\value{
\code{0} TODO
}
\description{
TODO
}
\author{
Rawane Samb
}
\keyword{internal}

