% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjmcmcMethod.R
\name{mergeRDSFiles}
\alias{mergeRDSFiles}
\title{Merge nucleosome information for selected RDS files.}
\usage{
mergeRDSFiles(RDSFiles)
}
\arguments{
\item{RDSFiles}{a \code{array}, the
names of all RDS used to merge nucleosome information. The files must
contain R object of \code{class} "rjmcmcNucleosomes" or
"rjmcmcNucleosomesMerge".}
}
\value{
a \code{list} of \code{class} "rjmcmcNucleosomesMerge" containing:
\itemize{
    \item k a \code{integer}, the number of nucleosomes.
    \item mu a \code{vector} of \code{numeric} of length
\code{k}, the positions of the nucleosomes.
    \item sigmaf a \code{vector} of \code{numeric} of length
\code{k}, the variance of the forward reads for each nucleosome.
    \item sigmar a \code{vector} of \code{numeric} of length
\code{k}, the variance of the reverse reads for each nucleosome.
    \item delta a \code{vector} of \code{numeric} of length
\code{k}, the distance between the maxima of the forward and
reverse reads position densities for each nucleosome.
}
}
\description{
Merge nucleosome information present in RDS files into one
object of \code{class} "rjmcmcNucleosomesMerge".
}
\examples{

## Use RDS files present in the RJMCMC package
RDSFiles <- dir(system.file("extdata", package = "RJMCMC"),
full.names = TRUE, pattern = "*rds")

## Merge nucleosomes info from RDS files present in directory
result <- mergeRDSFiles(RDSFiles)

## Print the number and the position of the nucleosomes
result$k
result$mu

## Class of the output object
class(result)


}
\author{
Pascal Belleau, Astrid Deschenes
}

