% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjmcmcMethodsIntern.R
\name{mhMove}
\alias{mhMove}
\title{Metropolis-Hastings move when more than one nucleosomes are present}
\usage{
mhMove(paramValues, kValue, muValue, sigmafValue, sigmarValue, deltaValue,
  wValue, dfValue, aValue, dimValue)
}
\arguments{
\item{paramValues}{a \code{list} containing:
\itemize{
    \item startPSF a \code{vector} of positive \code{integer}, the
start position of all the forward reads.
    \item startPSR a \code{vector} of positive \code{integer}, the
start position of all the reverse reads.
    \item kmax a \code{integer} the maximum number of nucleosomes allowed.
    \item lambda a positive \code{numeric}, the theorical mean
of the Poisson distribution.
    \item minReads a \code{integer}, the minimum
number of reads in a potential canditate region.
    \item y a \code{vector} of \code{numeric}, the position of all reads (
including forward and reverse reads).
    \item nr a \code{integer}, the number of reverse reads.
    \item nf a \code{integer}, the number of forward reads.
    \item nbrReads a \code{integer}, the total number of reads (including
forward and reverse reads).
    \item zeta \code{integer}, the length of a nucleosome.
    \item deltamin a \code{integer}, the minimum distance between the maxima
of the forward and reverse reads position densities for each nucleosome.
    \item deltamax a \code{integer}, the maximum distance between the maxima
of the forward and reverse reads position densities for each nucleosome.
    \item minReadPos a \code{numeric}, the minimum position of the reads.
    \item maxReadPos a \code{numeric}, the maximum position of the reads.
}}

\item{kValue}{a \code{integer}, the number of nucleosomes.}

\item{muValue}{a \code{vector} of \code{numeric} of length
\code{kValue}, the positions of the nucleosomes.}

\item{sigmafValue}{a \code{vector} of \code{numeric} of length
\code{kValue}, the variance of the forward reads for each nucleosome.}

\item{sigmarValue}{a \code{vector} of \code{numeric} of length
\code{kValue}, the variance of the reverse reads for each nucleosome.}

\item{deltaValue}{a \code{vector} of \code{numeric} of length
\code{kValue}, the distance between the maxima of the forward and
reverse reads position densities for each nucleosome.}

\item{wValue}{a \code{vector} of positive \code{numerical} of length
\code{kValue}, the weight for each nucleosome.}

\item{dfValue}{a \code{vector} of positive \code{numerical} of length
\code{kValue}, the degrees of freedom for each nucleosome.}

\item{aValue}{a \code{vector} of positive \code{numerical} of length
\code{kValue + 1}, the positions, on the chromosome, delimiting the regions
of the reads associated with each nucleosome.}

\item{dimValue}{a \code{vector} of positive \code{integer} of length
\code{kValue}, the number of reads associated to each nucleosome.}
}
\value{
a \code{list} containing:
\itemize{
    \item k a \code{integer}, the updated number of nucleosomes.
    \item mu a \code{vector} of \code{numeric} of length
\code{k}, the updated positions of the nucleosomes.
    \item sigmaf a \code{vector} of \code{numeric} of length
\code{k}, the updated variance of the forward reads for each nucleosome.
    \item sigmar a \code{vector} of \code{numeric} of length
\code{k}, the updated variance of the reverse reads for each nucleosome.
    \item delta a \code{vector} of \code{numeric} of length
\code{k}, the updated distance between the maxima of the forward and
reverse reads position densities for each nucleosome.
    \item w a \code{vector} of positive \code{numerical} of length
\code{k}, the updated weight for each nucleosome.
    \item df a \code{vector} of positive \code{numerical} of length
\code{k}, the updated degrees of freedom for each nucleosome.
    \item a a \code{vector} of positive \code{integer} of length
\code{k + 1}, the updated positions, on the chromosome, delimiting the
regions of the reads associated with each nucleosome.
    \item dim a \code{vector} of positive \code{integer} of length
\code{k}, the updated number of reads associated to each nucleosome.
    \item rho a \code{numeric}, the acceptance probability.
}
}
\description{
Attempt to randomly change the position of one nucleosome.
This move is given by the usual Metropolis-Hastings move, see for instance
Tierney (1994).
}
\examples{

## Load dataset
data(reads_demo)

## Create a list containing the mandatory parameters
paramList <- list(kmax = 5L, nf = length(reads_demo$readsForward),
nr = length(reads_demo$readsReverse),
nbrReads = length(reads_demo$readsForward) + length(reads_demo$readsReverse),
y = sort(c(reads_demo$readsForward, reads_demo$readsReverse)),
startPSF = reads_demo$readsForward,
startPSR = reads_demo$readsReverse, lambda = 2,
zeta = 147, deltamin = 142, deltamax = 152,
minReadPos = min(c(reads_demo$readsReverse, reads_demo$readsForward)),
maxReadPos = max(c(reads_demo$readsReverse, reads_demo$readsForward)))

## Initial position of the nucleosomes
muPosition <- c(72800, 73000)
muPosition

## Metropolis-Hastings move
result <- RJMCMC:::mhMove(paramValues = paramList, kValue = 2L,
muValue = muPosition, sigmafValue = c(100, 120), sigmarValue = c(120, 100),
deltaValue = c(200, 210), wValue = c(0.5, 0.5), dfValue = c(3, 4),
aValue = c(min(c(reads_demo$readsReverse, reads_demo$readsForward)), 73150,
max(c(reads_demo$readsReverse, reads_demo$readsForward))),
dimValue = c(length(reads_demo$readsForward),
length(reads_demo$readsReverse)))

## Final position of the nucleosomes
result$mu[1:2]

}
\author{
Rawane Samb, Pascal Belleau, Astrid Deschenes
}
\keyword{internal}

