% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rjmcmcMethodsIntern.R
\name{normal.mixture}
\alias{normal.mixture}
\title{Normal Mixture Model}
\usage{
normal.mixture(i, k, weight, mu, sigma)
}
\arguments{
\item{i}{a \code{vector} TODO}

\item{k}{a positive \code{integer} value, the number of nucleosomes in the
analyzed region.}

\item{weight}{a \code{vector} TODO}

\item{mu}{a \code{vector} of positive \code{integer} values, the positions
of all the nucleosomes in the analyzed region.}

\item{sigma}{a \code{vector} TODO}
}
\value{
\code{0} TODO
}
\description{
TODO
}
\author{
Rawane Samb
}
\keyword{internal}

