% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjmcmcMethod.R
\name{postTreatment}
\alias{postTreatment}
\title{A post treatment function to merge closely positioned nucleosomes
, from the same chromosome,
identified by the \code{\link{rjmcmc}} function..}
\usage{
postTreatment(startPosForwardReads, startPosReverseReads, resultRJMCMC, nbBase,
  chrLength)
}
\arguments{
\item{startPosForwardReads}{a \code{vector} of \code{numeric}, the
start position of all the forward reads.}

\item{startPosReverseReads}{a \code{vector} of \code{numeric}, the
start position of all the reverse reads. Beware that the start position of
a reverse read is always higher that the end positition.}

\item{resultRJMCMC}{TODO}

\item{nbBase}{a positive \code{numeric} or a positive \code{integer}
indicating TODO. The numeric will be treated as an integer.}

\item{chrLength}{a positive \code{numeric} or a positive \code{integer}
indicating the lenght of the current chromosome. The length of the
chromosome is used to ensure that the consensus positions are all
located inside the chromosome.}
}
\value{
a \code{array} of \code{numeric}, the updated values of the
nucleosome positions.
}
\description{
A helper function which merges closely positioned nucleosomes
to rectify the over splitting and provide a more conservative approach.
Beware that each chromosome must be treated separatly.
}
\examples{

## Loading dataset
data(reads_demo)

## Nucleosome positioning, running both merge and split functions
result <- rjmcmc(startPosForwardReads = reads_demo$readsForward,
         startPosReverseReads = reads_demo$readsReverse,
         nbrIterations = 1000, lambda = 2, kMax = 30,
         minInterval = 146, maxInterval = 292, minReads = 5)

##
postResult <- postTreatment(startPosForwardReads = reads_demo$readsForward,
         startPosReverseReads = reads_demo$readsReverse, result, 1, 73500)

postResult

}
\author{
Pascal Belleau, Astrid Deschenes
}

