% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.rjmcmcNucleosomes.R
\name{print.rjmcmcNucleosomes}
\alias{print.rjmcmcNucleosomes}
\title{Formated output of predicted nucleosomes}
\usage{
\method{print}{rjmcmcNucleosomes}(x, ...)
}
\arguments{
\item{x}{the output object from \code{rjmcmc} function to be printed}

\item{\ldots}{arguments passed to or from other methods}
}
\description{
Generated a formated output of a list marked as
an \code{rjmcmcNucleosomes} class
}
\examples{

## Loading dataset
data(RJMCMC_result)

print(RJMCMC_result)

}
\author{
Astrid Deschenes
}

