% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rjmcmcMethodsIntern.R
\name{priormu}
\alias{priormu}
\title{TODO}
\usage{
priormu(mu, y)
}
\arguments{
\item{sample}{a \code{vector} TODO}
}
\value{
\code{0} TODO
}
\description{
TODO
}
\author{
Rawane Samb
}
\keyword{internal}

