% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjmcmcMethodsIntern.R
\name{priorMuDensity}
\alias{priorMuDensity}
\title{Prior density of \eqn{mu}}
\usage{
priorMuDensity(mu, readPositions)
}
\arguments{
\item{mu}{a \code{vector} of positive \code{integer} containing the
positions of all nucleosomes.}

\item{readPositions}{a \code{vector} of positive \code{integer}
corresponding to the
positions of all reads, including forward and reverse strands. The
values insinde \code{readPositions} must be sorted.}
}
\value{
a \code{numeric}, the exact prior density of \code{mu} given the
number of nucleosomes.
}
\description{
Computes the prior density of \eqn{mu} conditionally to
the number of nucleosomes.

For more information on the calculation of the prior density of \eqn{mu},
see Proposotion 1 and equation (11) of the cited article.
}
\examples{

## Sorted vector of read positions, including forward and reverse
readPositions <- c(9909L, 9928L, 9935L, 26603L, 26616L, 26632L, 26636L,
             26640L, 44900L, 44902L, 44909L,  44910L, 44910L, 44918L,
             44924L, 44931L, 44935L, 44942L, 44946L)

## Position of the group of nucleosomes
mu <- c(10000L, 26700L, 45000L)

## Calculation of the exact prior density of mu
density <- RJMCMC:::priorMuDensity(mu, readPositions)

}
\author{
Rawane Samb, Astrid Deschenes
}
\references{
Samb R., Khadraoui K., Belleau P., Deschenes A., Lakhal L.
and Droit A. Using
informative Multinomial-Dirichlet prior in a t-mixture with
reversible jump estimation of nucleosome positions for genome-wide
profiling. Statistical Applications in Genetics and Molecular Biology.
Accepted (2015).
}
\keyword{internal}

