% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rjmcmc.R
\docType{data}
\name{reads_demo}
\alias{reads_demo}
\title{Forward reads and reverse reads (for demo purpose).}
\format{A \code{list} containing:
\itemize{
    \item \code{yf} a \code{vector} of \code{numeric}, the positions of the
forward reads.
    \item \code{yr} a \code{vector} of \code{numeric}, the positions of the
reverse reads.
}}
\source{
The Encyclopedia of DNA Elements (ENCODE) (DCC accession:
ENCFF002CFN)
}
\usage{
data(reads_demo)
}
\description{
Forward reads and reverse reads (for demo purpose).
}
\examples{
## Loading dataset
data(reads_demo)

## Nucleosome positioning
RJMCMC(yf = reads_demo$readsForward, yr = reads_demo$readsReverse,
     nbrIterations = 100, lambda = 3, kmax = 30,
     minInterval = 146, maxInterval = 292, minReads = 5)
}
\seealso{
\itemize{
    \item \code{\link{RJMCMC}} {for doing nucleosomes positioning.}
}
}
\keyword{datasets}

