% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rjmcmc.R
\docType{data}
\name{reads_demo_02}
\alias{reads_demo_02}
\title{Forward reads and reverse reads following a Normal distribution
(for demo purpose).}
\format{A \code{list} containing:
\itemize{
    \item \code{readsFoward} a \code{vector} of non-negative \code{numeric},
    the start positions of the forward reads.
    \item \code{readsReverse} a \code{vector} of non-negative
    \code{numeric}, the start
    positions of the reverse reads. Beware that the start position of
    a reverse read is always higher that the end positition.
}}
\usage{
data(reads_demo_02)
}
\description{
A group of froward and reverse reads that can be used to test the
\code{RJMCMC} function.
}
\examples{
## Loading dataset
data(reads_demo_02)

## Nucleosome positioning
RJMCMC(startPosForwardReads = reads_demo_02$readsForward,
         startPosReverseReads = reads_demo_02$readsReverse,
         nbrIterations = 150, lambda = 3, kmax = 30,
         minInterval = 144, maxInterval = 290, minReads = 6)
}
\seealso{
\itemize{
    \item \code{\link{RJMCMC}} {for doing nucleosomes positioning.}
}
}
\keyword{datasets}

