% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RJMCMC.R
\docType{package}
\name{RJMCMC-package}
\alias{RJMCMC-package}
\alias{RJMCMC}
\title{RJMCMC: Bayesian hierarchical model for genome-wide profiling of nucleosome positions based on high-throughput short-read data (MNase-Seq data)}
\description{
This package does nucleosome positioning using informative
Multinomial-Dirichlet prior in a t-mixture with reversible jump
estimation of nucleosome positions for genome-wide profiling.
}
\seealso{
\itemize{
    \item \code{\link{rjmcmc}} { for profiling of nucleosome positions}
    \item \code{\link{postTreatment}} { for merging closely positioned
    nucleosomes}
}
}
\author{
Rawane Samb,
Khader Khadraoui,
Pascal Belleau,
Astrid Deschênes,
Lajmi Lakhal and
Arnaud Droit

Maintainer:
Astrid Deschenes <astrid-louise.deschenes@crchudequebec.ulaval.ca>
}
\keyword{package}
