% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RJMCMC.R
\docType{data}
\name{RJMCMC_result}
\alias{RJMCMC_result}
\title{Nucleosomes obtained by running RJMCMC function
(for demo purpose).}
\format{A \code{list} of \code{class} "rjmcmcNucleosomes" containing:
\itemize{
\item \code{call} the matched call.
\item \code{K} a \code{vector} of \code{integer}, the estimation of the
number of the nucleosomes for each iteration.
\item \code{k} a \code{integer}, the final estimation of the number
of nucleosomes.
\item \code{mu} a \code{vector} of \code{numeric} of length
\code{k}, the positions of the nucleosomes.
\item \code{sigmaf} a \code{vector} of \code{numeric} of length
\code{k}, the variance of the forward reads for each nucleosome.
\item \code{sigmar} a \code{vector} of \code{numeric} of length
\code{k}, the variance of the reverse reads for each nucleosome.
\item \code{delta} a \code{vector} of \code{numeric} of length
\code{k}, the distance between the maxima of the forward and reverse reads
position densities for each nucleosome.
\item \code{df} a \code{vector} of \code{numeric} of length
\code{k}, the degrees of freedom for each nucleosome.
\item \code{w} a \code{vector} of positive \code{numerical} of length
\code{k}, the weight for each nucleosome. The sum of all \code{w} values
must be equal to \code{1}.
\item \code{qmu} a \code{matrix} of \code{numerical} with a number of rows
of \code{k}, the 2.5\% and 97.5\% quantiles of each \code{mu}.
\item \code{qsigmaf} a \code{matrix} of \code{numerical} with a number of
rows of \code{k}, the 2.5\% and 97.5\% quantiles of the variance of the
forward reads for each nucleosome.
\item \code{qsigmar} a \code{matrix} of \code{numerical} with a number of
rows of \code{k}, the 2.5\% and 97.5\% quantiles of the variance the
reverse reads for each nucleosome.
\item \code{qdelta} a \code{matrix} of \code{numerical} with a number of
rows of \code{k}, the 2.5\% and 97.5\% quantiles of the distance between
the maxima of the forward and reverse reads
position densities for each nucleosome.
\item \code{qdf} a \code{matrix} of \code{numerical} with a number of
rows of \code{k}, the 2.5\% and 97.5\% quantiles of the degrees of freedom
for each nucleosome.
\item \code{qw} a \code{matrix} of \code{numerical} with a number of rows
of \code{k}, the 2.5\% and 97.5\% quantiles of the weight for each
nucleosome.
}}
\usage{
data(RJMCMC_result)
}
\description{
A \code{list} of \code{class}
"rjmcmcNucleosomes" which contains the information about the
detected nucleosomes.
}
\examples{

## Loading dataset
data(RJMCMC_result)
data(reads_demo)

## Results before post-treatment
RJMCMC_result$mu

## Post-treatment function which merged closely positioned nucleosomes
postResult <- postTreatment(startPosForwardReads = reads_demo$readsForward,
startPosReverseReads = reads_demo$readsReverse, RJMCMC_result, 74, 73500)

## Results after post-treatment
postResult

}
\seealso{
\itemize{
    \item \code{\link{rjmcmc}} {for profiling of nucleosome positions}
}
}
\keyword{datasets}
