% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rjmcmcMethodsIntern.R
\name{splitNucleosome}
\alias{splitNucleosome}
\title{Spliting a nucleosomal region into two regions}
\usage{
splitNucleosome(yf, yr, y, liste, ecartmin, ecartmax, minReads)
}
\arguments{
\item{yf}{TODO}

\item{yr}{TODO}

\item{y}{TODO}

\item{liste}{TODO}

\item{ecartmin}{TODO}

\item{ecartmax}{TODO}

\item{minReads}{TODO}
}
\value{
\code{0} TODO
}
\description{
TODO
}
\author{
Rawane Samb
}
\keyword{internal}

