% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rjmcmcMethodsIntern.R
\name{student.mixture}
\alias{student.mixture}
\title{Student Mixture Model}
\usage{
student.mixture(i, k, w, mu, sigma, dfr)
}
\arguments{
\item{i}{a \code{vector} TODO}

\item{k}{a positive \code{integer} value, the number of nucleosomes in a
region.}

\item{w}{a \code{vector} weight for the nucleosome occupancy.}

\item{mu}{a \code{vector} mean of the Student-t distribution.}

\item{sigma}{a \code{vector} TODO}

\item{dfr}{a \code{vector} of \code{numeric} containing the degree
     of freedom.}
}
\value{
a \code{numerical}, the value generated from a Student Mixture
distribution.
}
\description{
TODO
}
\author{
Rawane Samb
}
\keyword{internal}

