% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rjmcmcMethodsIntern.R
\name{tnormale}
\alias{tnormale}
\title{Truncated random deviate from a normal distribution}
\usage{
tnormale(mu, sigma, a, b)
}
\arguments{
\item{mu}{a \code{numerical} The mean value of the normal distribution.}

\item{sigma}{a \code{vector} TODO}

\item{a}{a \code{numeric} The inferior boundary of the range
     in which the deviate value must be located. The
     deviate value has to be superior to \code{a}.}

\item{b}{a \code{numeric} The superior boundary of the range
     in which the deviate value must be located. The
     deviate value has to be inferior to \code{b}.}
}
\value{
\code{0} TODO
}
\description{
Generate a random deviate value from a normal
   distribution. The value must be included inside a
   specified range.
}
\author{
Rawane Samb
}
\keyword{internal}

