% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rjmcmcMethodsIntern.R
\name{validateParameters}
\alias{validateParameters}
\title{Parameters validation for the \code{\link{RJMCMC}}
function}
\usage{
validateParameters(yf, yr, niter, kmax, lambda, minInterval, maxInterval,
  minReads)
}
\arguments{
\item{yf}{a \code{vector} of positive \code{integer}, the positions of all
the forward reads.}

\item{yr}{a \code{vector} of positive \code{integer}, the positions of all
the reverse reads.}

\item{niter}{a positive \code{integer} or \code{numeric}, the number of
iterations. Non-integer values
of \code{niter} will be casted to \code{integer} and truncated towards
zero.}

\item{kmax}{a positive \code{integer} or \code{numeric}, the maximum number
of nucleosomes per region. Non-integer values
of \code{kmax} will be casted to \code{integer} and truncated towards zero.}

\item{lambda}{TODO}

\item{maxInterval}{a \code{numeric}, the maximum distance between two
nucleosomes.}

\item{minReads}{a positive \code{integer} or \code{numeric}, the minimum
number of reads in a potential canditate region. Non-integer values
of \code{niter} will be casted to \code{integer} and truncated towards
zero.}

\item{minInteval}{a \code{numeric}, the minimum distance between two
nucleosomes.}
}
\value{
\code{0} indicating that all parameters validations have been
successful.
}
\description{
Validation of all parameters needed by the public
\code{\link{RJMCMC}} function.
}
\author{
Astrid Louise Deschenes
}
\keyword{internal}

