% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rjmcmcMethodsIntern.R
\name{birthMoveK1}
\alias{birthMoveK1}
\title{Birth move in the case that only one nucleosome is present}
\usage{
birthMoveK1(paramValues, kValue, muValue, sigmafValue, sigmarValue, deltaValue,
  wValue, dlValue, aValue, dimValue)
}
\arguments{
\item{paramValues}{a \code{list} containing:
\itemize{
    \item startPSF a \code{vector} of \code{numeric}, the
start position of all the forward reads.
    \item startPSR a \code{vector} of \code{numeric}, the
start position of all the reverse reads.
    \item kmax a positive \code{integer} or \code{numeric}, the maximum
number of nucleosomes per region.
    \item lambda a positive \code{numeric}, the theorical mean
of the Poisson distribution.
    \item minReads a positive \code{integer} or \code{numeric}, the minimum
number of reads in a potential canditate region.
    \item y TODO
    \item nr TODO
    \item nf TODO
    \item nbrReads TODO
    \item zeta TODO
    \item deltamin TODO
    \item detlamax TODO
    \item minReadPos TODO
    \item maxReadPos TODO
}}

\item{kValue}{a \code{integer}, the number of nucleosomes.}

\item{muValue}{a \code{vector} of \code{numeric} of length
\code{kValue}, the positions of the nucleosomes.}

\item{sigmafValue}{a \code{vector} of \code{numeric} of length
\code{kValue}, the variance of the forward reads for each nucleosome.}

\item{sigmarValue}{a \code{vector} of \code{numeric} of length
\code{kValue}, the variance of the reverse reads for each nucleosome.}

\item{deltaValue}{}

\item{wValue}{a \code{vector} of positive \code{numerical} of length
\code{kValue}, the weight for each nucleosome. The sum of all \code{wValue}
values must be equal to \code{1}.}

\item{dlValue}{}

\item{aValue}{}

\item{dimValue}{}

\item{paramValues}{= list(startPSF =  startPosForwardReads, startPSR = startPosReverseReads
   , kmax = kmax, lambda = lambda, minReads = minReads
   , y = y, nr = nr, nf =nf, nbrReads = nbrReads , zeta = zeta
   , deltamin = deltamin, deltamax = deltamax
   , minReadPos= minReadPos, maxReadPos = maxReadPos )
   kValue, muValue, sigmafValue, sigmarValue, deltaValue, wValue, dlValue, aValue, dimValue}
}
\value{
list( k=0L, mu=numeric(parValue$kmax)
   , sigmaf=numeric(parValue$kmax), sigmar=numeric(parValue$kmax)
   , delta=numeric(parValue$kmax), w=numeric(parValue$kmax)
   , dl=numeric(parValue$kmax), a=numeric(parValue$kmax)
   , dim=numeric(parValue$kmax), rho=0)
}
\description{
Attempt to add a new nucleosome in the case that only one
nucleosome is present, \code{case k = 1}.
}
\examples{
birthMoveK1(paramValues , kValue, muValue, sigmafValue, sigmarValue,
deltaValue, wValue, dlValue, aValue, dimValue )
}
\author{
Rawane Samb, Pascal Belleau, Astrid Deschenes
}

