% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rjmcmcMethodsIntern.R
\name{birthMoveK1}
\alias{birthMoveK1}
\title{Birth move in the case that only one nucleosome is present}
\usage{
birthMoveK1(paramValues, kValue, muValue, sigmafValue, sigmarValue, deltaValue,
  wValue, dlValue, aValue, dimValue)
}
\arguments{
\item{paramValues}{= list(startPSF =  startPosForwardReads, startPSR = startPosReverseReads
   , kmax = kmax, lambda = lambda, minReads = minReads
   , y = y, nr = nr, nf =nf, nbrReads = nbrReads , zeta = zeta
   , deltamin = deltamin, deltamax = deltamax
   , minReadPos= minReadPos, maxReadPos = maxReadPos )}

\item{kValue}{}

\item{muValue}{a \code{vector} of \code{numeric} of length
\code{kValue}, the positions of the nucleosomes.}

\item{sigmafValue}{a \code{vector} of \code{numeric} of length
\code{kValue}, the variance of the forward reads for each nucleosome.}

\item{sigmarValue}{a \code{vector} of \code{numeric} of length
\code{kValue}, the variance of the reverse reads for each nucleosome.}

\item{deltaValue}{a \code{vector} of \code{numeric} of length
\code{kValue}, the distance between the maxima of the forward and
reverse reads position densities for each nucleosome.}

\item{wValue}{a \code{vector} of positive \code{numerical} of length
\code{kValue}, the weight for each nucleosome.}

\item{dlValue}{TODO}

\item{aValue}{}

\item{dimValue}{}

\item{paramValues}{a \code{list} containing:
\itemize{
    \item startPSF a \code{vector} of positive \code{integer}, the
start position of all the forward reads.
    \item startPSR a \code{vector} of positive \code{integer}, the
start position of all the reverse reads.
    \item kmax a \code{integer} the maximum number of nucleosomes allowed.
    \item lambda TODO
    \item minReads TODO
    \item y TODO
    \item nr TODO
    \item nf TODO
    \item nbrReads TODO
    \item zeta TODO
    \item deltamin TODO
    \item deltamax TODO
    \item minReadPos TODO
    \item maxReadPos TODO
}}
}
\value{
list( k=0L, mu=numeric(paramValues$kmax)
   , sigmaf=numeric(paramValues$kmax), sigmar=numeric(paramValues$kmax)
   , delta=numeric(paramValues$kmax), w=numeric(paramValues$kmax)
   , dl=numeric(paramValues$kmax), a=numeric(paramValues$kmax)
   , dim=numeric(paramValues$kmax), rho=0)
}
\description{
Attempts to add a component for a new nucleosome in the
case that only one nucleosome is present \code{k = 1}
}
\examples{
birthMoveK1(paramValues, kValue, muValue, sigmafValue, sigmarValue,
deltaValue, wValue, dlValue, aValue, dimValue )
}
\author{
Rawane Samb, Pascal Belleau, Astrid Deschenes
}
\keyword{internal}

