% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rjmcmcMethodsIntern.R
\name{birthMoveK1}
\alias{birthMoveK1}
\title{birthMoveK1 birth move in the case k = 1}
\usage{
birthMoveK1(paramValues, kValue, muValue, sigmafValue, sigmarValue, deltaValue,
  wValue, dlValue, aValue, dimValue)
}
\arguments{
\item{paramValues}{= list(startPSF =  startPosForwardReads, startPSR = startPosReverseReads
   , kmax = kmax, lambda = lambda, minReads = minReads
   , y = y, nr = nr, nf =nf, nbrReads = nbrReads , zeta = zeta
   , deltamin = deltamin, deltamax = deltamax
   , minReadPos= minReadPos, maxReadPos = maxReadPos )
   kValue, muValue, sigmafValue, sigmarValue, deltaValue, wValue, dlValue, aValue, dimValue}
}
\value{
list( k=0L, mu=numeric(paramValues$kmax)
   , sigmaf=numeric(paramValues$kmax), sigmar=numeric(paramValues$kmax)
   , delta=numeric(paramValues$kmax), w=numeric(paramValues$kmax)
   , dl=numeric(paramValues$kmax), a=numeric(paramValues$kmax)
   , dim=numeric(paramValues$kmax), rho=0)
}
\description{
birth move
}
\examples{
birthMoveK1(paramValues , kValue, muValue, sigmafValue, sigmarValue, deltaValue, wValue, dlValue, aValue, dimValue )
}
\author{
Rawane Samb, Pascal Belleau, Astrid Deschenes
}
\keyword{internal}

