% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rjmcmcMethodsIntern.R
\name{birthMoveK1}
\alias{birthMoveK1}
\title{Birth move in the case that only one nucleosome is present}
\usage{
birthMoveK1(paramValues, kValue, muValue, sigmafValue, sigmarValue, deltaValue,
  wValue, dlValue, aValue, dimValue)
}
\arguments{
\item{paramValues}{= list(startPSF =  startPosForwardReads, startPSR = startPosReverseReads
   , kmax = kmax, lambda = lambda, minReads = minReads
   , y = y, nr = nr, nf =nf, nbrReads = nbrReads , zeta = zeta
   , deltamin = deltamin, deltamax = deltamax
   , minReadPos= minReadPos, maxReadPos = maxReadPos )
   kValue, muValue, sigmafValue, sigmarValue, deltaValue, wValue, dlValue, aValue, dimValue}
}
\value{
list( k=0L, mu=numeric(parValue$kmax)
   , sigmaf=numeric(parValue$kmax), sigmar=numeric(parValue$kmax)
   , delta=numeric(parValue$kmax), w=numeric(parValue$kmax)
   , dl=numeric(parValue$kmax), a=numeric(parValue$kmax)
   , dim=numeric(parValue$kmax), rho=0)
}
\description{
Attempt to add a new nucleosome in the case that only one
nucleosome is present, \code{case k = 1}.
}
\examples{
birthMoveK1(paramValues , kValue, muValue, sigmafValue, sigmarValue,
deltaValue, wValue, dlValue, aValue, dimValue )
}
\author{
Rawane Samb, Pascal Belleau, Astrid Deschenes
}

