% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rjmcmcMethodsIntern.R
\name{Bk}
\alias{Bk}
\title{Birth Submove Probability}
\usage{
Bk(k, lambda, kMax)
}
\arguments{
\item{k}{a \code{numerical} used as the lambda
     value in the truncated Poisson distribution.}

\item{lambda}{a \code{numerical} used as the lambda
     value in the truncated Poisson distribution.}

\item{kMax}{a \code{numeric} indicating the maximum
     value authorized for the \code{k} parameter. When
     \code{k} is equal or superior to \code{kMax}, the
     returned value is \code{0}.}
}
\value{
a \code{numerical} value. The value \code{0} when
     \code{k} is equal or superior to \code{kMax} or
     when \code{k} is equal to \code{1}.
}
\description{
Calculation of the birth submove
     probability using a truncated Poisson distribution.
}
\author{
Rawane Samb
}
\keyword{internal}

