% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rjmcmcMethodsIntern.R
\name{Bk}
\alias{Bk}
\title{Birth Submove Probability}
\usage{
Bk(k, lambda, kMax)
}
\arguments{
\item{k}{a positive \code{integer}, the number of
     nucleosomes.}

\item{lambda}{a \code{numeric}, the theorical mean
     of the Poisson distribution.}

\item{kMax}{a positive \code{integer}, the maximum number of nucleosomes
     authorized. When
     \code{k} is equal or superior to \code{kMax}, the
     returned value is \code{0}.}
}
\value{
a \code{numeric} value. The value \code{0} when
     \code{k} is equal or superior to \code{kMax} or
     when \code{k} is equal to \code{1}.
}
\description{
Calculation of the birth submove
     probability of adding a new nucleosome using a
     truncated Poisson distribution.
}
\examples{
## Return the birth submove probability
rjmcmc:::Bk(k = 14L, lambda = 2.4, kMax = 22L)

## Zero is returned when k = 1
rjmcmc:::Bk(k = 1L, lambda = 3.4, kMax = 20L)

## Zerio is returned when k is superior to kMax
rjmcmc:::Bk(k = 31L, lambda = 4.4, kMax = 30L)
}
\author{
Rawane Samb
}
\keyword{internal}

