% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rjmcmcMethodsIntern.R
\name{Dk}
\alias{Dk}
\title{TODO}
\usage{
Dk(k, lambda, kmax)
}
\arguments{
\item{k}{a \code{numerical} TODO}

\item{lambda}{a \code{vector} TODO}

\item{kmax}{a \code{numeric} TODO}
}
\value{
\code{0} TODO
}
\description{
TODO
}
\author{
Rawane Samb
}
\keyword{internal}

