% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjmcmcMethodsIntern.R
\name{mergeAllRDSFiles}
\alias{mergeAllRDSFiles}
\title{Merge nucleosome information}
\usage{
mergeAllRDSFiles(arrayOfFiles)
}
\arguments{
\item{arrayOfFiles}{a \code{array}, the name of each file that must be
used to merge nucleosome information.}
}
\value{
a \code{list} of \code{class} "rjmcmcNucleosomesMerge" containing:
\itemize{
    \item k a \code{integer}, the number of nucleosomes.
    \item mu a \code{vector} of \code{numeric} of length
\code{k}, the positions of the nucleosomes.
    \item sigmaf a \code{vector} of \code{numeric} of length
\code{k}, the variance of the forward reads for each nucleosome.
    \item sigmar a \code{vector} of \code{numeric} of length
\code{k}, the variance of the reverse reads for each nucleosome.
    \item delta a \code{vector} of \code{numeric} of length
\code{k}, the distance between the maxima of the forward and
reverse reads position densities for each nucleosome.
}
}
\description{
Merge nucleosome information present in multiple RDS files.
}
\examples{

## Loading two files containing nucleosomes informations for two sections of
## the same chromosome
file_100 <- dir(system.file("extdata", package = "RJMCMC"),
                pattern = "yeastRes_Chr1_Seg_100.rds",
                full.names = TRUE)

file_101 <- dir(system.file("extdata", package = "RJMCMC"),
                pattern = "yeastRes_Chr1_Seg_101.rds",
                full.names = TRUE)

## Merging nucleosomes informations from the two files
result <- RJMCMC:::mergeAllRDSFiles(c(file_100, file_101))

}
\author{
Pascal Belleau, Astrid Deschenes
}
\keyword{internal}
