% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rjmcmcMethodsIntern.R
\name{mergeNucleosomes}
\alias{mergeNucleosomes}
\title{Merging two nucleosomal regions}
\usage{
mergeNucleosomes(yf, yr, y, liste, minInterval, maxInterval, minReads)
}
\arguments{
\item{yf}{a \code{vector} of positive \code{integer}
corresponding to the
positions of all forward reads. The
values inside \code{yf} must be sorted.}

\item{yr}{a \code{vector} of positive \code{integer}
corresponding to the
positions of all reverse reads. The
values inside \code{yf} must be sorted.}

\item{y}{TODO}

\item{liste}{TODO}

\item{minInterval}{a \code{numeric}, the minimum distance between two
nucleosomes.}

\item{maxInterval}{a \code{numeric}, the maximum distance between two
nucleosomes.}

\item{minReads}{a positive \code{integer}, the minimum
number of reads in a potential canditate region.}
}
\value{
\code{0} TODO
}
\description{
TODO
}
\author{
Rawane Samb
}
\keyword{internal}

