% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rjmcmcMethodsIntern.R
\name{mergeNucleosomes}
\alias{mergeNucleosomes}
\title{Merging two nucleosomal regions}
\usage{
mergeNucleosomes(yf, yr, y, liste, minInterval, maxInterval, minReads)
}
\arguments{
\item{yf}{a \code{vector} of positive \code{numeric}
corresponding to the
positions of all forward reads. The
values inside \code{yf} must be sorted.}

\item{yr}{a \code{vector} of positive \code{numeric}
corresponding to the
positions of all reverse reads. The
values inside \code{yr} must be sorted.}

\item{y}{a \code{vector} of positive \code{numeric}
corresponding to the
positions of all forward and reverse reads. The
values inside \code{y} must be sorted.}

\item{liste}{a \code{list} containing:
\itemize{
    \item k a \code{integer}, the number of nucleosomes.
    \item mu a \code{vector} of \code{numeric}, the positions of
the nucleosomes.
    \item sigmaf a \code{vector} of \code{numeric} of length
\code{k}, the variance of the forward reads for each nucleosome.
    \item sigmar a \code{vector} of \code{numeric} of length
\code{k}, the variance of the reverse reads for each nucleosome.
    \item delta TODO
    \item dl TODO
    \item w a \code{vector} of positive \code{numerical} of length
\code{k}, the weight for each nucleosome. The sum of all \code{w} values
must be equal to \code{1}.
}}

\item{minInterval}{a \code{numeric}, the minimum distance between two
nucleosomes.}

\item{maxInterval}{a \code{numeric}, the maximum distance between two
nucleosomes.}

\item{minReads}{a positive \code{integer}, the minimum
number of reads in a potential canditate region.}
}
\value{
a \code{list} containing the updated values:
\itemize{
    \item k a \code{integer}, the number of nucleosomes.
    \item mu a \code{vector} of \code{numeric}, the positions of
the nucleosomes.
    \item sigmaf a \code{vector} of \code{numeric} of length
\code{k}, the variance of the forward reads for each nucleosome.
    \item sigmar a \code{vector} of \code{numeric} of length
\code{k}, the variance of the reverse reads for each nucleosome.
    \item delta TODO
    \item dl TODO
    \item w a \code{vector} of positive \code{numerical} of length
\code{k}, the weight for each nucleosome. The sum of all \code{w} values
must be equal to \code{1}.
}
}
\description{
Merging two nucleosomal regions into one region with respect of
the minimal and maximal intervals allowed.
}
\author{
Rawane Samb, Astrid Deschenes
}
\keyword{internal}

