% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rjmcmcMethodsIntern.R
\name{normal.mixture}
\alias{normal.mixture}
\title{TODO}
\usage{
normal.mixture(i, k, w, mu, sigma)
}
\arguments{
\item{i}{a \code{vector} TODO}

\item{k}{a \code{vector} TODO}

\item{w}{a \code{vector} TODO}

\item{mu}{a \code{vector} TODO}

\item{sigma}{a \code{vector} TODO}
}
\value{
\code{0} TODO
}
\description{
TODO
}
\author{
Rawane Samb
}
\keyword{internal}

