% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjmcmcMethodsIntern.R
\name{postMerge}
\alias{postMerge}
\title{A internal post treatment function to merge closely positioned
nucleosomes, from the same chromosome,
identified by the \code{\link{rjmcmc}} function.}
\usage{
postMerge(startPosForwardReads, startPosReverseReads, resultRJMCMC, nbBase,
  chrLength)
}
\arguments{
\item{startPosReverseReads}{a \code{vector} of \code{numeric}, the
start position of all the reverse reads. Beware that the start position of
a reverse read is always higher that the end positition.}

\item{resultRJMCMC}{an object of class 'rjmcmcNucleosomes' or
'rjmcmcNucleosomesMerge' containing informations about nucleosomes.}

\item{nbBase}{a positive \code{numeric} or a positive \code{integer}
indicating TODO. The numeric will be treated as an integer.}

\item{chrLength}{a positive \code{numeric} or a positive \code{integer}
indicating the lenght of the current chromosome. The length of the
chromosome is used to ensure that the consensus positions are all
located inside the chromosome.}

\item{startPosFrowardReads}{a \code{vector} of \code{numeric}, the
start position of all the forward reads.}
}
\value{
a \code{array} of \code{numeric}, the updated values of the
nucleosome positions.
}
\description{
A internal helper function which merges closely positioned
nucleosomes to rectify the over splitting and provide a more conservative
approach. Beware that each chromosome must be treated separatly.
}
\author{
Pascal Belleau, Astrid Deschenes
}
\keyword{internal}

