% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rjmcmcMethodsIntern.R
\name{priormu}
\alias{priormu}
\title{Prior expectation of \eqn{mu}}
\usage{
priormu(mu, y)
}
\arguments{
\item{mu}{a \code{vector} TODO}

\item{y}{a \code{vector} TODO}
}
\value{
\code{0} TODO
}
\description{
TODO
}
\author{
Rawane Samb
}
\keyword{internal}

