% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rjmcmcMethodsIntern.R
\name{priormu}
\alias{priormu}
\title{Prior density of \eqn{mu}}
\usage{
priormu(mu, reads)
}
\arguments{
\item{mu}{a \code{vector} of positive \code{integer} containing the
     positions of all nucleosomes.}

\item{reads}{a \code{vector} of \code{TODO} corresponding to the read
     data, including forward and reverse strands.}
}
\value{
the exact prior density of \code{\mu} given the
     number of nucleosomes.
}
\description{
Computes the prior density of \eqn{mu} conditionally to
 the number of nucleosomes.

 For more information on the calculation of the prior density of \equ{mu},
 see Proposotion 1 of the cited article.
}
\author{
Rawane Samb, Astrid Louise Deschenes
}
\references{
Samb R., Khadraoui K., Lakhal L., Belleau P. and Droit A. Using
     informative Multinomial-Dirichlet prior in a t-mixture with
     reversible jump estimation of nucleosome positions for genome-wide
     profiling. Submitted (2015).
}
\keyword{internal}

