% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RJMCMC.R
\docType{package}
\name{RJMCMC-package}
\alias{RJMCMC}
\alias{RJMCMC-package}
\title{RJMCMC: Nucleosome Positioning}
\description{
This package does nucleosome positioning using informative
Multinomial-Dirichlet prior in a t-mixture with reversible jump
estimation of nucleosome positions for genome-wide profiling.
}
\author{
Rawane Samb,
Khader Khadraoui,
Pascal Belleau,
Astrid Deschenes,
Lajmi Lakhal and
Arnaud Droit

Maintainer:
Astrid Deschenes <astrid-louise.deschenes@crchudequebec.ulaval.ca>
}
\seealso{
\itemize{
    \item \code{\link{RJMCMC}} {for doing nucleosomes positioning.}
}
}
\keyword{package}

