% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rjmcmc.R
\docType{package}
\name{rjmcmc-package}
\alias{rjmcmc}
\alias{rjmcmc-package}
\title{rjmcmc: Nucleosome Positioning}
\description{
This package does nucleosome positioning using informative
Multinomial-Dirichlet prior in a t-mixture with reversible jump
estimation of nucleosome positions for genome-wide profiling.
}
\author{
Rawane Samb,
Astrid Louise Deschenes,
Khader Khadraoui,
Lajmi Lakhal,
Pascal Belleau and
Arnaud Droit

Maintainer:
Astrid Louise Deschenes <astrid-louise.deschenes@crchudequebec.ulaval.ca>
}
\seealso{
\itemize{
    \item \code{\link{RJMCMC}} {for doing nucleosomes positioning.}
}
}
\keyword{package}

