% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rjmcmcMethod.R
\name{RJMCMC}
\alias{RJMCMC}
\title{Nucleosome positioning mapping}
\usage{
RJMCMC(yf, yr, niter, kmax, lambda, ecartmin, ecartmax, minReads)
}
\arguments{
\item{yf}{a \code{vector} of positive \code{integer}, the positions of all
the forward reads.}

\item{yr}{a \code{vector} of positive \code{integer}, the positions of all
the reverse reads.}

\item{niter}{a \code{numeric} number of iterations.}

\item{kmax}{a positive \code{integer}, the maximum number of nucleosomes
per region.}

\item{lambda}{}

\item{ecartmin}{a \code{numeric}, the minimum distance between two
nucleosomes.}

\item{ecartmax}{a \code{numeric}, the maximum distance between two
nucleosomes.}

\item{minReads}{a \code{numeric} minimum number of reads in a potential
canditate region.}
}
\value{
\code{0} TODO
}
\description{
Use of a fully Bayesian hierarchical model for genome-wide
profiling of nucleosome positions based on high-throughput short-read
data (MNase-Seq data).
}
\author{
Rawane Samb
}

