% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjmcmcMethod.R
\name{rjmcmc}
\alias{rjmcmc}
\title{Nucleosome positioning mapping}
\usage{
rjmcmc(startPosForwardReads, startPosReverseReads, nbrIterations, kMax,
  lambda = 3, minInterval, maxInterval, minReads = 5,
  adaptIterationsToReads = TRUE)
}
\arguments{
\item{startPosForwardReads}{a \code{vector} of \code{numeric}, the
start position of all the forward reads.}

\item{startPosReverseReads}{a \code{vector} of \code{numeric}, the
start position of all the reverse reads. Beware that the start position of
a reverse read is always higher that the end positition.}

\item{nbrIterations}{a positive \code{integer} or \code{numeric}, the
number of iterations. Non-integer values of
\code{nbrIterations} will be casted to \code{integer} and truncated towards
zero.}

\item{kMax}{a positive \code{integer} or \code{numeric}, the maximum number
of nucleosomes per region. Non-integer values
of \code{kMax} will be casted to \code{integer} and truncated towards zero.}

\item{lambda}{a positive \code{numeric}, the theorical mean
of the Poisson distribution. Default: 3.}

\item{minInterval}{a \code{numeric}, the minimum distance between two
nucleosomes.}

\item{maxInterval}{a \code{numeric}, the maximum distance between two
nucleosomes.}

\item{minReads}{a positive \code{integer} or \code{numeric}, the minimum
number of reads in a potential canditate region. Non-integer values
of \code{minReads} will be casted to \code{integer} and truncated towards
zero. Default: 5.}

\item{adaptIterationsToReads}{a \code{logical} indicating if the number
of iterations must be modified in function of the number of reads.
Default: \code{TRUE}.}
}
\value{
a \code{list} of \code{class} "rjmcmcNucleosomes" containing:
\itemize{
\item \code{call} the matched call.
\item \code{K} a \code{vector} of \code{integer}, the estimation of the
number of the nucleosomes for each iteration.
\item \code{k} a \code{integer}, the final estimation of the number
of nucleosomes.
\item \code{mu} a \code{vector} of \code{numeric} of length
\code{k}, the positions of the nucleosomes.
\item \code{sigmaf} a \code{vector} of \code{numeric} of length
\code{k}, the variance of the forward reads for each nucleosome.
\item \code{sigmar} a \code{vector} of \code{numeric} of length
\code{k}, the variance of the reverse reads for each nucleosome.
\item \code{delta} a \code{vector} of \code{numeric} of length
\code{k}, the distance between the maxima of the forward and reverse reads
position densities for each nucleosome.
\item \code{df} a \code{vector} of \code{numeric} of length
\code{k}, the degrees of freedom for each nucleosome.
\item \code{w} a \code{vector} of positive \code{numerical} of length
\code{k}, the weight for each nucleosome. The sum of all \code{w} values
must be equal to \code{1}.
\item \code{qmu} a \code{matrix} of \code{numerical} with a number of rows
of \code{k}, the 2.5\% and 97.5\% quantiles of each \code{mu}.
\item \code{qsigmaf} a \code{matrix} of \code{numerical} with a number of
rows of \code{k}, the 2.5\% and 97.5\% quantiles of the variance of the
forward reads for each nucleosome.
\item \code{qsigmar} a \code{matrix} of \code{numerical} with a number of
rows of \code{k}, the 2.5\% and 97.5\% quantiles of the variance the
reverse reads for each nucleosome.
\item \code{qdelta} a \code{matrix} of \code{numerical} with a number of
rows of \code{k}, the 2.5\% and 97.5\% quantiles of the distance between
the maxima of the forward and reverse reads
position densities for each nucleosome.
\item \code{qdf} a \code{matrix} of \code{numerical} with a number of
rows of \code{k}, the 2.5\% and 97.5\% quantiles of the degrees of freedom
for each nucleosome.
\item \code{qw} a \code{matrix} of \code{numerical} with a number of rows
of \code{k}, the 2.5\% and 97.5\% quantiles of the weight for each
nucleosome.
}
}
\description{
Use of a fully Bayesian hierarchical model for chromosome-wide
profiling of nucleosome positions based on high-throughput short-read
data (MNase-Seq data). Beware that for a genome-wide profiling, each
chromosome must be treated separatly.
}
\examples{

## Loading dataset
data(reads_demo)

## Nucleosome positioning, running both merge and split functions
result <- rjmcmc(startPosForwardReads = reads_demo$readsForward,
         startPosReverseReads = reads_demo$readsReverse,
         nbrIterations = 1000, lambda = 2, kMax = 30,
         minInterval = 146, maxInterval = 292, minReads = 5)

## Print the final estimation of the number of nucleosomes
result$k

## Print the position of nucleosomes
result$mu

}
\author{
Rawane Samb, Pascal Belleau, Astrid Deschênes
}

