% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rjmcmcMethod.R
\name{RJMCMC}
\alias{RJMCMC}
\title{TODO}
\usage{
RJMCMC(yf, yr, niter, kmax, lambda, ecartmin, ecartmax, minReads)
}
\arguments{
\item{yf}{a \code{numerical} TODO}

\item{yr}{a \code{vector} TODO}

\item{niter}{a \code{numeric} TODO}

\item{kmax}{a \code{numeric} TODO}

\item{lambda}{a \code{numeric} TODO}

\item{ecartmin}{a \code{numeric} TODO}

\item{ecartmax}{a \code{numeric} TODO}

\item{minReads}{a \code{numeric} TODO}
}
\value{
\code{0} TODO
}
\description{
TODO
}
\author{
Rawane Samb
}

