% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rjmcmcMethod.R
\name{RJMCMC}
\alias{RJMCMC}
\title{Nucleosome positioning mapping}
\usage{
RJMCMC(yf, yr, nbrIterations, kmax, lambda, minInterval, maxInterval, minReads)
}
\arguments{
\item{yf}{a \code{vector} of positive \code{integer}, the positions of all
the forward reads.}

\item{yr}{a \code{vector} of positive \code{integer}, the positions of all
the reverse reads.}

\item{nbrIterations}{a positive \code{integer} or \code{numeric}, the
number of iterations. Non-integer values of
\code{nbrIterations} will be casted to \code{integer} and truncated towards
zero.}

\item{kmax}{a positive \code{integer} or \code{numeric}, the maximum number
of nucleosomes per region. Non-integer values
of \code{kmax} will be casted to \code{integer} and truncated towards zero.}

\item{lambda}{a positive \code{numeric}, the theorical mean
of the Poisson distribution.}

\item{minInterval}{a \code{numeric}, the minimum distance between two
nucleosomes.}

\item{maxInterval}{a \code{numeric}, the maximum distance between two
nucleosomes.}

\item{minReads}{a positive \code{integer} or \code{numeric}, the minimum
number of reads in a potential canditate region. Non-integer values
of \code{minReads} will be casted to \code{integer} and truncated towards
zero.}
}
\value{
an \code{list} of \code{class} "rjmcmcNucleosomes" containing :
\itemize{
\item \code{call} the matched call.
\item \code{K} TODO
\item \code{k} a \code{integer}, the number of nucleosomes.
\item \code{mu} a \code{vector} of \code{numeric}, the positions of
the nucleosomes.
\item \code{sigmaf} TODO
\item \code{sigmar} TODO
\item \code{delta} TODO
\item \code{dl} TODO
\item \code{w} TODO
\item \code{qmu} TODO
\item \code{qsigmaf} TODO
\item \code{qsigmar} TODO
\item \code{qdelta} TODO
\item \code{qdl} TODO
\item \code{qw} TODO
}
}
\description{
Use of a fully Bayesian hierarchical model for genome-wide
profiling of nucleosome positions based on high-throughput short-read
data (MNase-Seq data).
}
\author{
Rawane Samb
}

