% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rjmcmcMethod.R
\name{RJMCMC}
\alias{RJMCMC}
\title{Nucleosome positioning mapping}
\usage{
RJMCMC(startPosForwardReads, startPosReverseReads, nbrIterations, kmax, lambda,
  minInterval, maxInterval, minReads)
}
\arguments{
\item{startPosForwardReads}{a \code{vector} of \code{numeric}, the
start position of all the forward reads.}

\item{startPosReverseReads}{a \code{vector} of \code{numeric}, the
start position of all the reverse reads. Beware that the start position of
a reverse read is always higher that the end positition.}

\item{nbrIterations}{a positive \code{integer} or \code{numeric}, the
number of iterations. Non-integer values of
\code{nbrIterations} will be casted to \code{integer} and truncated towards
zero.}

\item{kmax}{a positive \code{integer} or \code{numeric}, the maximum number
of nucleosomes per region. Non-integer values
of \code{kmax} will be casted to \code{integer} and truncated towards zero.}

\item{lambda}{a positive \code{numeric}, the theorical mean
of the Poisson distribution.}

\item{minInterval}{a \code{numeric}, the minimum distance between two
nucleosomes.}

\item{maxInterval}{a \code{numeric}, the maximum distance between two
nucleosomes.}

\item{minReads}{a positive \code{integer} or \code{numeric}, the minimum
number of reads in a potential canditate region. Non-integer values
of \code{minReads} will be casted to \code{integer} and truncated towards
zero.}
}
\value{
a \code{list} of \code{class} "rjmcmcNucleosomes" containing :
\itemize{
\item \code{call} the matched call.
\item \code{K} a \code{vector} of \code{integer}, the number of
the nucleosomes for each iteration.
\item \code{k} a \code{integer}, the number of nucleosomes.
\item \code{mu} a \code{vector} of \code{numeric} of length
\code{k}, the positions of the nucleosomes.
\item \code{sigmaf} a \code{vector} of \code{numeric} of length
\code{k}, the variance of the forward reads for each nucleosome.
\item \code{sigmar} a \code{vector} of \code{numeric} of length
\code{k}, the variance of the reverse reads for each nucleosome.
\item \code{delta} a \code{vector} of \code{numeric} of length
\code{k}, the distance between the maxima of the forward and reverse reads
position densities for each nucleosome.
\item \code{dl} TODO
\item \code{w} a \code{vector} of positive \code{numerical} of length
\code{k}, the weight for each nucleosome. The sum of all \code{w} values
must be equal to \code{1}.
\item \code{qmu} a \code{matrix} of \code{numerical} with a number of rows
of \code{k}, the 2.5\% and 97.5\% quantiles of each \code{mu}.
\item \code{qsigmaf} a \code{matrix} of \code{numerical} with a number of
rows of \code{k}, the 2.5\% and 97.5\% quantiles of the variance of the
forward reads for each nucleosome.
\item \code{qsigmar} a \code{matrix} of \code{numerical} with a number of
rows of \code{k}, the 2.5\% and 97.5\% quantiles of the variance the
reverse reads for each nucleosome.
\item \code{qdelta} a \code{matrix} of \code{numerical} with a number of
rows of \code{k}, the 2.5\% and 97.5\% quantiles of the distance between
the maxima of the forward and reverse reads
position densities for each nucleosome.
\item \code{qdl} TODO
\item \code{qw} a \code{matrix} of \code{numerical} with a number of rows
of \code{k}, the 2.5\% and 97.5\% quantiles of each \code{w}.
}
}
\description{
Use of a fully Bayesian hierarchical model for genome-wide
profiling of nucleosome positions based on high-throughput short-read
data (MNase-Seq data).
}
\examples{
## Loading dataset
data(reads_demo)

## Nucleosome positioning
result <- RJMCMC(startPosForwardReads = reads_demo$readsForward,
         startPosReverseReads = reads_demo$readsReverse,
         nbrIterations = 1000, lambda = 2, kmax = 30,
         minInterval = 146, maxInterval = 292, minReads = 5)

## Print the number of nucleosomes
result$k

## Print the position of nucleosomes
result$mu
}
\author{
Rawane Samb
}

