% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rjmcmcMethodsIntern.R
\name{splitNucleosome}
\alias{splitNucleosome}
\title{Spliting a nucleosomal region into two regions}
\usage{
splitNucleosome(yf, yr, y, liste, minInterval, maxInterval, minReads)
}
\arguments{
\item{yf}{TODO}

\item{yr}{TODO}

\item{y}{TODO}

\item{liste}{TODO}

\item{maxInterval}{a \code{numeric}, the maximum distance between two
nucleosomes.}

\item{minReads}{a positive \code{integer}, the minimum
number of reads in a potential canditate region.}

\item{minInteval}{a \code{numeric}, the minimum distance between two
nucleosomes.}
}
\value{
\code{0} TODO
}
\description{
TODO
}
\author{
Rawane Samb
}
\keyword{internal}

