% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rjmcmcMethodsIntern.R
\name{tnormale}
\alias{tnormale}
\title{Truncated random deviate from a normal distribution}
\usage{
tnormale(mu, sigma, a, b)
}
\arguments{
\item{mu}{a \code{numerical} The mean of the normal distribution.}

\item{sigma}{a \code{vector} The variance of the normal distribution.}

\item{a}{a \code{numeric} The inferior boundary of the range
     in which the output value must be located. The
     output value has to be superior to \code{a}.}

\item{b}{a \code{numeric} The superior boundary of the range
     in which the output value must be located. The
     output value has to be inferior to \code{b}.}
}
\value{
a \code{numeric} superior to \code{a} and inferior
     to \code{b}.
}
\description{
Generate a random deviate value from a normal
   distribution. The returned value is included inside a
   specified range ]a,b[ specified by user. The mean and
   variance of the normal distribution is also specified by
   user.
}
\author{
Rawane Samb
}
\keyword{internal}

