% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjmcmcMethodsIntern.R
\name{tnormale}
\alias{tnormale}
\title{Random deviate from a truncated normal distribution}
\usage{
tnormale(mu, sigma, minValue, maxValue)
}
\arguments{
\item{mu}{a \code{numeric} value, the mean of the normal distribution.}

\item{sigma}{a non-negative \code{numeric}, the variance of the normal
distribution.}

\item{minValue}{a \code{numeric} value, the inferior boundary of the
range in which the output value must be located. The returned value has to
be superior to \code{minValue}.}

\item{maxValue}{a \code{numeric} value, the superior boundary of the range
in which the output value must be located. The returned value has to be
inferior to \code{maxValue}.}
}
\value{
a \code{numeric} which is superior to \code{minValue} and inferior
to \code{maxValue}.
}
\description{
Generate a random deviate value from a normal distribution.
The returned value is included inside a specified range
]\code{minValue},\code{maxValue}[
specified by user. The mean and variance of the normal distribution is
also specified by user.
}
\examples{

## Set seed to replicate results
set.seed(3333)

## Return a value, between 1000 and 3000, generated froma a normal
## distribution with a mean of 2000 and a variance of 30.
RJMCMC:::tnormale(2000, 30, 1000, 30000)

}
\author{
Rawane Samb
}
\keyword{internal}

