% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjmcmcMethodsIntern.R
\name{tnormale}
\alias{tnormale}
\title{Random deviate from a truncated normal distribution}
\usage{
tnormale(mu, sigma, minValue, maxValue)
}
\arguments{
\item{mu}{a \code{numeric} value, the mean of the normal distribution.}

\item{sigma}{a non-negative \code{numeric}, the variance of the normal
distribution.}

\item{minValue}{a \code{numeric} value, the inferior boundary of the
range in which the output value must be located. The returned value has to be
superior to \code{minValue}.}

\item{maxValue}{a \code{numeric} value, the superior boundary of the range
in which the output value must be located. The returned value has to be
inferior to \code{maxValue}.}
}
\value{
a \code{numeric} which is superior to \code{minValue} and inferior
to \code{maxValue}.
}
\description{
Generate a random deviate value from a normal distribution.
The returned value is included inside a specified range ]minValue,maxValue[
specified by user. The mean and variance of the normal distribution is
also specified by user.
}
\author{
Rawane Samb
}
\keyword{internal}

