% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rjmcmcMethodsIntern.R
\name{tnormale}
\alias{tnormale}
\title{Truncated random deviate from a normal distribution}
\usage{
tnormale(mu, sigma, minValue, maxValue)
}
\arguments{
\item{mu}{a \code{numerical} mean of the normal distribution.}

\item{sigma}{a \code{numerical} variance of the normal distribution.}

\item{minValue}{a \code{numeric} inferior boundary of the range
     in which the output value must be located. The
     output value has to be superior to \code{minValue}.}

\item{maxValue}{a \code{numeric} superior boundary of the range
     in which the output value must be located. The
     output value has to be inferior to \code{maxValue}.}
}
\value{
a \code{numeric} superior to \code{minValue} and inferior
     to \code{maxValue}.
}
\description{
Generate a random deviate value from a normal
   distribution. The returned value is included inside a
   specified range ]minValue,maxValue[ specified by user. The mean and
   variance of the normal distribution is also specified by
   user.
}
\author{
Rawane Samb
}
\keyword{internal}

