% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjmcmcMethodsIntern.R
\name{validateDirectoryParameters}
\alias{validateDirectoryParameters}
\title{Parameters validation for the
\code{\link{mergeAllRDSFilesFromDirectory}} function}
\usage{
validateDirectoryParameters(directory)
}
\arguments{
\item{directory}{a \code{character}, the
name of the directory (relative or absolute path) containing RDS files.}
}
\value{
\code{0} indicating that all parameters validations have been
successful.
}
\description{
Validation of all parameters needed by the public
\code{\link{mergeAllRDSFilesFromDirectory}} function.
}
\examples{

## Load an existing directory
directory <- system.file("extdata", package = "RJMCMC")

## Testing using a real directory
RJMCMC:::validateDirectoryParameters(directory)

}
\author{
Astrid Deschenes
}
\keyword{internal}
