% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rjmcmcMethodsIntern.R
\name{validateParameters}
\alias{validateParameters}
\title{Parameters validation for the \code{\link{RJMCMC}}
function}
\usage{
validateParameters(yf, yr, nbrIterations, kmax, lambda, minInterval,
  maxInterval, minReads)
}
\arguments{
\item{yf}{a \code{vector} of positive \code{integer}, the positions of all
the forward reads.}

\item{yr}{a \code{vector} of positive \code{integer}, the positions of all
the reverse reads.}

\item{nbrIterations}{a positive \code{integer} or \code{numeric}, the
number of iterations. Non-integer values of
\code{nbrIterations} will be casted to \code{integer} and truncated towards
zero. The maximum value of \code{nbrIterations} is \code{100000}.}

\item{kmax}{a positive \code{integer} or \code{numeric}, the maximum number
of nucleosomes per region. Non-integer values
of \code{kmax} will be casted to \code{integer} and truncated towards zero.}

\item{lambda}{a positive \code{numeric}, the theorical mean
of the Poisson distribution.}

\item{minInterval}{a \code{numeric}, the minimum distance between two
nucleosomes.}

\item{maxInterval}{a \code{numeric}, the maximum distance between two
nucleosomes.}

\item{minReads}{a positive \code{integer} or \code{numeric}, the minimum
number of reads in a potential canditate region. Non-integer values
of \code{minReads} will be casted to \code{integer} and truncated towards
zero.}
}
\value{
\code{0} indicating that all parameters validations have been
successful.
}
\description{
Validation of all parameters needed by the public
\code{\link{RJMCMC}} function.
}
\author{
Astrid Louise Deschenes
}
\keyword{internal}

