% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rjmcmcMethodsIntern.R
\name{validateParameters}
\alias{validateParameters}
\title{Parameters validation for the \code{\link{rjmcmc}}
function}
\usage{
validateParameters(startPosForwardReads, startPosReverseReads, nbrIterations,
  kMax, lambda, minInterval, maxInterval, minReads, adaptIterationsToReads)
}
\arguments{
\item{startPosForwardReads}{a \code{vector} of positive \code{integer}, the
start position of all the forward reads.}

\item{startPosReverseReads}{a \code{vector} of positive \code{integer}, the
positions of all the reverse reads. Beware that the start position of
a reverse read is always higher that the end positition.}

\item{nbrIterations}{a positive \code{integer} or \code{numeric}, the
number of iterations. Non-integer values of
\code{nbrIterations} will be casted to \code{integer} and truncated towards
zero.}

\item{kMax}{a positive \code{integer} or \code{numeric}, the maximum number
of nucleosomes per region. Non-integer values
of \code{kMax} will be casted to \code{integer} and truncated towards zero.}

\item{lambda}{a positive \code{numeric}, the theorical mean
of the Poisson distribution.}

\item{minInterval}{a \code{numeric}, the minimum distance between two
nucleosomes.}

\item{maxInterval}{a \code{numeric}, the maximum distance between two
nucleosomes.}

\item{minReads}{a positive \code{integer} or \code{numeric}, the minimum
number of reads in a potential canditate region. Non-integer values
of \code{minReads} will be casted to \code{integer} and truncated towards
zero.}

\item{adaptIterationsToReads}{a \code{logical} indicating if the number
of iterations must be modified in function of the number of reads.}
}
\value{
\code{0} indicating that all parameters validations have been
successful.
}
\description{
Validation of all parameters needed by the public
\code{\link{rjmcmc}} function.
}
\author{
Astrid Deschenes
}
\keyword{internal}

