% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjmcmcMethodsIntern.R
\name{validatePrepMergeParameters}
\alias{validatePrepMergeParameters}
\title{Parameters validation for the \code{\link{postMerge}} function.}
\usage{
validatePrepMergeParameters(startPosForwardReads, startPosReverseReads,
  resultRJMCMC, extendingSize, chrLength)
}
\arguments{
\item{startPosForwardReads}{a \code{vector} of positive \code{integer}, the
start position of all the forward reads.}

\item{startPosReverseReads}{a \code{vector} of positive \code{integer}, the
positions of all the reverse reads. Beware that the start position of
a reverse read is always higher that the end position.}

\item{resultRJMCMC}{an object of \code{class}
"rjmcmcNucleosomes" or "rjmcmcNucleosomesMerge" that contain information
 about nucleosome positioning for an entire chromosome.}

\item{extendingSize}{a positive \code{numeric} or a positive \code{integer}
indicating the size of the consensus region used to group closeley
positioned nucleosomes.The minimum size of the consensus region is equal to
twice the value of the \code{extendingSize} parameter. The numeric will
be treated as an integer.}

\item{chrLength}{a positive \code{numeric} or a positive \code{integer}
indicating the lenght of the current chromosome. The length of the
chromosome is used to ensure that the consensus positions are all
located inside the chromosome.}
}
\value{
\code{0} indicating that all parameters validations have been
successful.
}
\description{
Validation of all parameters needed by the public
\code{\link{postMerge}} function.
}
\examples{

## Load dataset containing forward and reverse reads
data(reads_demo)

## Load dataset containing nucleosome information
file_002 <- dir(system.file("extdata", package = "RJMCMC"),
pattern = "yeastRes_Chr1_Seg_002.rds", full.names = TRUE)
nucleosome_info <- readRDS(file_002)

## Parameters validation
RJMCMC:::validatePrepMergeParameters(startPosForwardReads =
reads_demo$readsForward, startPosReverseReads = reads_demo$readsReverse,
resultRJMCMC = nucleosome_info, extendingSize = 74, chrLength = 10000000)

}
\author{
Astrid Deschenes
}
\keyword{internal}

